+++ cmd/gateway_http_wiring.go
@@ -278,6 +278,7 @@ func (d *gatewayDeps) wireHTTPHandlersOnServer(
 	// V3: Knowledge Vault document API
 	if d.pgStores != nil && d.pgStores.Vault != nil {
 		vh := httpapi.NewVaultHandler(d.pgStores.Vault, d.pgStores.Teams, d.workspace, d.domainBus, d.pgStores.Agents, d.pgStores.Teams)
+		vh.SetContactStore(d.pgStores.Contacts)
 		vh.SetEnrichProgress(d.enrichProgress)
 		vh.SetEnrichWorker(d.enrichWorker)
 		d.server.SetVaultHandler(vh)